review: changes part 1
- remove stubs
- compact Commands together, cfg's when diffrences
- update ci (no chip selection is a compile error)
- update st25r39-disco to use feature
