+++ ci.sh
@@ -17,12 +17,8 @@ cargo build --release --manifest-path rnfc-fm175xx/Cargo.toml --features ''
 cargo build --release --manifest-path rnfc-fm175xx/Cargo.toml --features 'defmt'
 cargo build --release --manifest-path rnfc-fm175xx/Cargo.toml --features 'log'
 
-echo "no-features"
-cargo build --release --manifest-path rnfc-st25r39/Cargo.toml --features ''
-echo "st25r3916"
 cargo build --release --manifest-path rnfc-st25r39/Cargo.toml --features 'st25r3916, defmt'
 cargo build --release --manifest-path rnfc-st25r39/Cargo.toml --features 'st25r3916, log'
-echo "st25r3911b"
 cargo build --release --manifest-path rnfc-st25r39/Cargo.toml --features 'st25r3911b, defmt'
 cargo build --release --manifest-path rnfc-st25r39/Cargo.toml --features 'st25r3911b, log'
